Deru Dataset

A subset of data from Déru et al. 2020 containing hologenomic data from 750 pigs fed a conventional diet. Genotypes are accessible thanks to the "population" attribute.

Usage

Deru

Format

A matrix with 780 rows (individuals) and 1845 taxa, where each element represents the abundance of a specific taxon in a specific individual. Additionnally, there is an attribute "population" wich provides the genotypes of each individual, encoded as 0, 1 ,2.

Source

Déru et al. for microbiota data and Déru et al. 2020 for genotypes data.

Examples

# Access the dataset
data("Deru")

# View the first few rows of the taxa matrix
head(Deru[1:5,1:5])
#>   OTU1 OTU2 OTU6793 OTU3 OTU4
#> 1   30  593       4  630  414
#> 2  254  275      62 1131  446
#> 3  181  487     164 1472  656
#> 4  469  665      45  640  328
#> 5  771  519      21  758  347

# Access the population attribute to view genotypes
population <- attr(Deru, "population")