Assign all taxa to a cluster, eventually under genetic control using hclust

Assign all taxa to a cluster, eventually under genetic control using hclust

Calibration of genetic effect from founder population data

Formatting data from file paths to base population object

assign_taxa(founder_object, taxa_g = 0.1)

gen_effect_calibration(
  founder_object,
  taxa_assign_g,
  correlation = 0.5,
  effect.size = c(seq(0.1, 1, 0.1)),
  plot = T
)

read_input_data(
  path_to_microbiome,
  path_to_pedmap,
  threshold = 0.05,
  ind_selected = NULL
)

Arguments

founder_object

Output of generate_founder() function

taxa_g

Percentage of taxa under genetic control, DEFAULT = 0.1

taxa_assign_g

Factor vector giving cluster assignment for all taxa, typical output of assign_taxa()

correlation

Correlation between taxa within the same cluster, value between 0 and 1, DEFAULT = 0.5

effect.size

Vector giving the size of genetic effect to try

plot

boolean, if plot generation is required, DEFAULT = TRUE

path_to_microbiome

String giving the path to count table file

path_to_pedmap

String giving the path and prefix to ped and map file

threshold

Threshold for rarefaction, DEFAULT = 0.05

ind_selected

Vector of string values with individuals to keep, have to match rownames of count table file, DEFAULT = NULL

Value

A vector with a length matching the total number of taxa with values from 0 to the number of taxa, 0 corresponding to the non under genetic control cluster

A data.frame with three columns, giving the Taxa ID, the effect.size and the corresponding heritability

A data.frame corresponding to (rarefied) microbiome with individuals in rows and taxa in columns. Genotypes data.frame is an attribute called "population" reachable using attr(output_name,"population")

Examples

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
datafile <- system.file("DeruPop.rds", package = "RITHMS")
DeruPop <- readRDS(datafile)
assign_taxa(founder_object = DeruPop,
            taxa_g = 0.2)
#>    [1] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>   [25] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>   [49] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>   [73] 0  0  0  0  0  0  0  0  0  0  0  0  0  1  0  0  0  0  0  0  0  0  0  0 
#>   [97] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [121] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [145] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [169] 0  0  0  0  0  0  0  0  0  2  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [193] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [217] 0  0  0  0  0  0  0  0  0  0  0  0  3  0  0  0  0  2  0  0  0  0  0  0 
#>  [241] 4  0  0  0  0  0  0  0  0  0  0  5  0  0  0  4  0  0  0  0  0  0  0  0 
#>  [265] 0  0  0  0  0  0  0  0  0  0  5  0  0  0  0  1  0  0  0  0  0  0  4  0 
#>  [289] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [313] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [337] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  4  0 
#>  [361] 0  0  0  0  0  0  0  0  0  0  5  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [385] 0  0  6  0  0  0  0  0  2  0  0  0  0  0  0  7  0  2  0  0  0  0  0  0 
#>  [409] 0  0  0  0  0  0  0  0  0  5  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [433] 0  0  0  0  0  0  0  0  5  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [457] 0  0  0  0  0  0  0  0  0  8  0  9  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [481] 0  0  0  0  0  0  0  4  0  0  0  0  0  0  0  0  0  0  0  0  0  0  9  0 
#>  [505] 6  0  0  3  0  0  0  0  0  0  5  0  0  0  0  0  0  0  0  0  0  0  0  0 
#>  [529] 0  7  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  4  0  0 
#>  [553] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  9  0  0  0  0  0 
#>  [577] 0  0  0  0  0  2  7  0  0  2  0  10 9  11 12 0  0  0  0  0  0  0  2  0 
#>  [601] 0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  11 0  0  13
#>  [625] 0  0  3  0  0  2  0  0  0  12 0  0  0  0  1  0  0  0  3  8  0  0  0  13
#>  [649] 11 9  0  0  0  0  0  6  0  0  0  0  0  0  8  0  0  3  0  0  0  4  0  0 
#>  [673] 0  0  0  0  0  0  0  6  0  6  0  0  0  0  2  9  0  0  8  0  0  0  0  0 
#>  [697] 0  4  0  0  0  0  0  3  0  4  0  0  0  1  0  0  0  0  0  0  0  0  0  0 
#>  [721] 0  10 0  0  0  0  0  0  0  0  0  0  0  4  0  0  12 0  0  0  0  0  10 0 
#>  [745] 0  0  0  11 0  0  0  0  0  2  0  0  0  0  0  0  13 0  0  0  0  0  0  0 
#>  [769] 0  9  0  0  0  1  0  0  0  6  0  0  0  0  0  0  0  0  0  0  14 0  0  0 
#>  [793] 10 0  0  0  7  0  0  0  4  0  15 0  0  3  0  0  16 0  4  0  0  0  0  0 
#>  [817] 0  0  6  0  0  0  0  0  16 17 0  0  0  15 0  0  0  0  8  0  0  0  2  0 
#>  [841] 0  0  0  7  0  0  0  0  6  0  0  0  0  0  0  6  9  0  0  0  1  0  0  0 
#>  [865] 12 0  0  0  0  11 0  0  0  0  17 0  0  16 0  0  0  0  0  0  9  9  0  0 
#>  [889] 0  0  0  0  0  0  0  0  0  0  9  0  0  0  0  0  16 0  11 18 0  0  0  19
#>  [913] 0  0  0  0  10 0  10 0  0  0  0  0  0  0  16 15 9  0  6  16 0  0  3  0 
#>  [937] 18 2  0  10 0  0  0  0  0  0  15 20 7  0  0  0  8  13 0  17 0  13 0  0 
#>  [961] 0  0  0  11 0  17 4  10 0  9  0  0  0  0  12 0  0  0  9  0  0  0  17 0 
#>  [985] 0  0  0  17 0  0  0  10 0  10 4  0  9  0  0  17 18 0  0  0  0  0  0  7 
#> [1009] 7  0  0  0  9  0  0  0  0  0  11 0  0  0  0  0  0  10 0  0  0  17 3  0 
#> [1033] 0  0  0  0  0  8  16 0  0  7  0  0  0  0  0  0  0  11 0  0  9  0  0  1 
#> [1057] 19 0  0  0  0  7  9  10 0  0  5  10 0  0  2  0  0  0  1  0  0  0  17 0 
#> [1081] 7  9  0  17 0  0  0  7  0  0  0  0  4  0  0  9  0  10 0  0  0  0  0  0 
#> [1105] 0  0  0  19 0  0  9  0  0  0  21 0  22 22 1  0  0  0  10 0  0  5  0  23
#> [1129] 10 0  22 0  0  0  0  0  0  21 0  0  9  0  0  0  0  1  0  10 0  21 4  0 
#> [1153] 0  12 16 0  0  0  0  0  0  0  18 0  0  0  0  0  0  4  0  15 8  4  0  0 
#> [1177] 0  0  0  0  0  0  0  0  8  12 22 16 0  5  0  0  0  15 0  21 10 4  0  0 
#> [1201] 0  0  0  0  0  18 0  13 0  21 20 0  23 14 0  0  0  0  0  0  16 0  24 9 
#> [1225] 9  1  20 0  15 16 3  0  0  0  0  10 0  0  18 0  0  0  10 19 15 8  0  0 
#> [1249] 14 0  4  3  0  0  0  0  0  19 16 16 0  0  1  5  0  0  0  0  0  0  0  0 
#> [1273] 0  0  0  0  0  14 0  0  10 0  0  0  0  0  0  7  0  19 0  0  0  18 0  10
#> [1297] 0  0  0  0  0  0  0  13 0  0  0  12 20 0  21 12 0  0  16 16 0  0  0  0 
#> [1321] 0  13 0  0  0  5  0  0  0  14 23 0  0  16 11 3  0  0  11 13 13 16 0  0 
#> [1345] 0  0  0  0  21 0  0  10 2  0  0  19 0  0  0  0  0  0  0  12 7  0  0  0 
#> [1369] 0  0  0  0  23 1  0  0  0  21 0  0  0  0  0  20 7  0  18 0  0  0  0  7 
#> [1393] 0  0  0  0  21 0  0  0  0  0  0  16 0  0  0  0  0  15 0  0  6  0  0  22
#> [1417] 0  0  0  0  19 0  0  0  24 12 0  0  0  0  17 0  0  0  0  0  0  7  7  0 
#> [1441] 0  19 0  0  0  0  22 0  7  0  7  24 0  24 0  0  0  0  0  14 18 0  0  0 
#> [1465] 0  0  10 0  0  0  0  0  0  0  0  0  0  13 0  0  0  0  0  24 0  0  0  13
#> [1489] 0  0  0  0  0  15 0  0  0  0  0  0  0  24 0  0  21 16 0  0  0  0  14 16
#> [1513] 0  0  0  0  15 15 24 0  0  0  23 0  0  0  0  0  18 0  0  12 0  0  0  22
#> [1537] 7  0  24 0  0  0  0  8  0  22 0  0  0  0  0  22 0  0  0  0  0  0  0  0 
#> [1561] 24 0  0  0  19 0  0  7  6  0  0  0  0  23 0  0  0  0  0  23 0  11 0  21
#> [1585] 0  0  0  14 0  5  0  0  0  0  0  0  0  19 0  0  21 0  0  0  0  0  21 0 
#> [1609] 0  0  23 0  17 24 0  21 0  13 0  0  0  0  0  0  0  0  0  14 0  17 18 0 
#> [1633] 0  0  7  0  0  14 0  7  0  14 14 0  21 20 0  2  20 21 0  20 0  18 0  0 
#> [1657] 0  21 0  0  19 0  0  21 0  0  0  0  0  0  0  0  0  0  0  0  23 0  0  0 
#> [1681] 0  0  0  0  19 0  0  0  14 0  23 0  0  0  0  0  0  0  0  0  0  0  0  15
#> [1705] 0  19 22 13 0  0  0  0  0  0  0  0  23 0  0  2  0  0  0  0  0  0  0  0 
#> [1729] 0  0  22 0  15 0  0  0  0  0  0  0  0  0  0  0  0  12 20 0  0  12 0  0 
#> [1753] 12 0  21 14 0  0  19 19 0  0  0  0  0  0  0  0  0  12 0  0  0  0  0  0 
#> [1777] 0  0  0  0  24 0  0  0  0  0  0  0  0  0  15 0  24 0  0  0  0  20 0  13
#> [1801] 0  0  0  2  11 21 0  20 0  0  0  0  0  24 0  0  0  0  15 0  0  13 0  0 
#> [1825] 0  0  22 0  0  0  0  0  15 0  0  0  21 0  0  0  0  0  0  0  0 
#> Levels: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24