Main function to generate transgenerational hologenomic data

Main function to generate transgenerational hologenomic data

holo_simu(
  h2,
  b2,
  founder_object,
  otu_g = 0.05,
  qtn_y = NULL,
  n_ind = NULL,
  n_gen = 5,
  n_clust = NULL,
  selection = F,
  size_selection_F = NULL,
  size_selection_M = NULL,
  selection_type = "GB",
  seed = 1234,
  verbose = T,
  lambda = 0.5,
  correlation = 0.5,
  mix.params = c(0.75, 0.25),
  mix.params.M = c(0.75, 0.25),
  effect.size = 0.1,
  noise.microbiome = 0.1,
  dir = F,
  ao = 25,
  thetaX = NULL,
  env_gen = NULL,
  w.param = c(0.5, 0.5)
)

Arguments

h2

direct heritability value, between 0 and 1

b2

microbiability value, between 0 and 1

founder_object

output of generate_founder() function

otu_g

percentage of taxa under genetic control, between 0 and 1, DEFAULT = 0.05

qtn_y

number of causal SNPs for the phenotypes

n_ind

number of individual per generation

n_gen

number of generation, DEFAULT = 5

n_clust

vector with taxa assignment, typical output of assign_taxa

selection

bool, if selection process needed, DEFAULT = FALSE

size_selection_F

percentage of female to select

size_selection_M

percentage of male to select

selection_type

mode of selection to be used, value in ("GB", "B", "G", "diversity", "div.GB"), DEFAULT = "GB"

seed

seed value for samplings in the function

verbose

bool, DEFAULT = T

lambda

proportion of microbiome of offspring coming from vertical transmission, DEFAULT = 0.5

correlation

Correlation between taxa within the same cluster, value between 0 and 1, DEFAULT = 0.5

mix.params

Vector of two values giving the weights for the regularisation of the base population microbiome. mix.params1 = weight for raw microbiome and mix.params2 = weight for mean microbiome. DEFAULT = c(0.75,0.25)

noise.microbiome

sd of microbiome noise, DEFAULT = 0.1

dir

bool

ao

for dirichlet sampling parameter

thetaX

if environmental effect

env_gen

vector of bool

w.param

in case div.GB selection mode is chosen

Value

A big list object with metada info such as beta matrix details and each generation at level 1. For each generation, the genotypes, the microbiomes, the phenotypes, the pedigree and the individuals selected can be reachable.